- 퀘스트 메뉴 슬라이드 바 수정 - 퀘스트 수락/완료 시 사운드 이펙트 추가

+++ 3D_RPG_Project/Assets/_3D RPG/Scripts/Enemy/EnemyUi.cs
@@ -44,6 +44,6 @@ public class EnemyUi : MonoBehaviour
         
 
         hpBar.value = _enemyStatus.GetCurrentHp() / (float)_enemyStatus.GetMaxHp();
-        _Ui.text = "Lv " + _enemyStatus.GetLevel() + " " + _enemyStatus.GetName() + " " + hpPercent + "%";
+        _Ui.text = "Lv " + _enemyStatus.GetLevel() + " " + _enemyStatus.GetName() + " " + (int)hpPercent + "%";
     }
 }
